Added visualization separation
